moves strong params into controller, deletes 5million files
