+++ app/controllers/authors_controller.rb
@@ -61,15 +61,15 @@ class AuthorsController < ApplicationController
     @author = Author.find(params[:id])
   end
 
-  def author_params
-    params.require(:author).permit(:name, :id, :sort_by)
-  end
-
   def sort_column(default = "name")
     params[:sort] ? params[:sort] : default
   end
   
   def sort_direction
     %w[asc desc].include?(params[:direction]) ? params[:direction] : "asc"
+  end
+
+  def author_params
+    params.require(:author).permit(:name, :id, :sort_by)
   end
 end
