--- app/controllers/concerns/keyword_helper.rb
@@ -1,5 +1,0 @@
-module KeywordHelper
-  def keyword_params
-    params.require(:keyword).permit(:name, :id)
-  end 
-end
