--- app/controllers/concerns/loan_helper.rb
@@ -1,9 +1,0 @@
-module LoanHelper
-  def loan_params
-    params.require(:loan).permit(:book_id, :user_id, :id)
-  end 
-
-  def find_loan
-    @loan = Loan.find(params[:id])
-  end
-end
