--- app/controllers/concerns/user_uploads_helper.rb
@@ -1,6 +1,0 @@
-module UserUploadsHelper
-
-  def user_uploads_params
-    params.require(:user_upload).permit(:file)
-  end
-end
