--- app/helpers/authors_helper.rb
@@ -1,2 +1,0 @@
-module AuthorsHelper
-end
