--- app/helpers/book_authors_helper.rb
@@ -1,2 +1,0 @@
-module BookAuthorsHelper
-end
