--- app/helpers/book_uploads_helper.rb
@@ -1,8 +1,0 @@
-module BookUploadsHelper
-
-  def book_uploads_params
-    params.require(:book_upload).permit(:file, :genre)
-  end
-  
-
-end
