--- app/helpers/books_helper.rb
@@ -1,5 +1,0 @@
-module BooksHelper
-  def book_params
-    params.require(:book).permit!
-  end
-end
