--- app/helpers/genres_helper.rb
@@ -1,2 +1,0 @@
-module GenresHelper
-end
