+++ app/helpers/users_helper.rb
@@ -1,11 +1,5 @@
 module UsersHelper
 
-  def user_params
-    params.require(:user).permit( :email, :password, :password_confirmation, :remember_me,
-                                  :notes, :name, :phone, :preferred_first_name, :do_not_lend,
-                                  :address, :city, :state, :zip, :identification, :username )
-  end
-
   def librarian_user?
     if current_user
       if current_user.has_role? :librarian
